move Theme class
