+++ app/src/main/java/com/github/monolabhituszi/portal/MainActivity.kt
@@ -9,7 +9,6 @@ import androidx.databinding.DataBindingUtil
 import androidx.navigation.fragment.NavHostFragment
 import androidx.navigation.ui.NavigationUI.setupWithNavController
 import com.github.monolabhituszi.portal.databinding.ActivityMainBinding
-import com.github.monolabhituszi.portal.ui.Theme
 import kotlinx.android.synthetic.main.activity_main.*
 
 class MainActivity : AppCompatActivity() {
